Fix PDOs sending remote frames instead of actual data
If a TPDO communication parameter has the RTR bit set, meaning remote
transmission requests are not allowed for this PDO, the PDO transmission
will send an RTR frame instead of the data when it's triggered, due to a
missing mask on the COB-ID when sending the frame.

Also audit other cases of calling os_channel_send() without masking and
as a result, replace hard-coded mask constants in SYNC sending (one of
which was incorrect!) with the appropriate macro.

Include the extended-frame bit in the mask to support 29-bit frames if
these should be configured in any COB-ID; it makes little sense to
include all 29 ID bits in the mask if the extended-frame flag is left
out.