Fix multiple data corruption in OD load
If the current OD had an object with a larger data size than when the
objects were stored, only the smaller part of the uint64_t would be read
into, but the actual setting of the object would use the entire, partly
uninitialized, value.

Also, in the pointer case the size was not checked against the actual size
of the target object, which would cause memory corruption if the object had
been shrunk since the values were stored.

Signed-off-by: Andreas Fritiofson <andreas.fritiofson@unjo.com>
Change-Id: I1e0fc0b0d6da4594ca8bcb61e01ab6ab5f42a551