Support access to integers of size not a power of two
Data types such as UNSIGNED48 was not supported, except via the access
function with the previous commit. Add all the non-power-of-two variants
of UNSIGNED and INTEGER to the OD get/set functions.

Note that defining an array of these objects using OD_ARRAY is broken,
because it produces unaligned accesses, that even if they may be
supported by the platform, will not necessarily be atomic.

Signed-off-by: Andreas Fritiofson <andreas.fritiofson@unjo.com>
Change-Id: Id7ee68cc50921b9b6705e95a5867b7f99eb319be