Fix corrupt OD load following a non-existent object
If an OD is loaded from a store that contains an object which has
subsequently been removed, it is necessary to skip the part of the data
which was stored for the object. Otherwise the data part will be
interpreted as the index, subindex and size of the next object to load,
possibly loading incorrect data or at least failing to load the
following objects.

Also skip the object if the current OD would not store it (OD_TRANSIENT)
and also abort the load if a size of zero is encountered in the store.

Signed-off-by: Andreas Fritiofson <andreas.fritiofson@unjo.com>