Fix SDO time-out on large segmented transfers
Any sufficiently large SDO transfer will time-out, because the time is
counted from the start of the transfer. Restart the timeout as long as
progress is made, for each transferred segment.

Also fill in the index and sub-index of the object in the abort frame
if the transfer times out locally, as per CiA 301.

Additionally, fix a flaw in the SegmentedTimeout test which used
synthetic timestamps but did not mock the system time, possibly causing
sporadic failure (or pass) if run at the wrong time.

Signed-off-by: Andreas Fritiofson <andreas.fritiofson@unjo.com>
Change-Id: Ic527f80cca1c3f72e26bebb0acd6d3b06800d546