Synchronize PDO triggering with the c-open main thread
If a call to trigger a PDO event is done followed by a second job,
there's no guarantee that the main thread will have a chance to run
and process the PDO job to completion before the common job
description struct is overwritten by the next request.

Use the semaphore via the completion callback as for all other jobs.

Change-Id: I49a59d9e3d69434a984186a12ea8f35b5289d62c