Pass the network handle to callback functions instead of cb_arg
It's currently impossible to reach the network handle from callbacks
that are called during co_init(), because even if the cb_arg is already
set up to point to a global object that is going to hold the network
handle, co_init() must return the handle before it can be stored there.

Add a function to retrieve the cb_arg opaque argument from the network
handle when needed. This is also useful to reach application specific
context in the object access functions.

Change-Id: Id14ca3b71dd064f0e7328053c77675d13152ae31