Split NMT Initialisation state into its three sub-states
Treat each sub-state as a separate step in the FSM processing, to expose
the sub-states to the application NMT callback, which may need to do
different things depending on whether the application or only the
communication is reset.

The previous STATE_INIT in an NMT callback now corresponds to
STATE_INIT_COMM, so add an alias to keep the behaviour of existing
callbacks.

Move the reset callback to the NMT command reception to avoid having to
disable it during power-on initialisation.

Change-Id: I6eb9745f55f5d85a6ae4cf09c72fcbec2df56577