Fix PDO mapping with OD_ARRAY entries
The second entry is returned from co_entry_find(), for array or record
objects, when that entry has the OD_ARRAY flag set. So for these objects
the mapped entry pointer in the PDO will always point to the same entry
regardless of the subindex of the mapping.

Therefore, the PDO packing and unpacking must not consider the subindex
field of the mapped entry when storing or fetching the value. Instead it
needs to look at the actual mapped subindex to access the correct data.

Change-Id: I983f25bdeb10de72f7522d7a2f912e0c885e59fb